Creating parse mod.
