+++ engine/src/conversion/parse/parse_bindgen.rs
@@ -12,11 +12,11 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use std::{collections::HashMap, collections::HashSet, fmt::Display};
+use std::{collections::HashMap, collections::HashSet};
 
 use crate::{
-    byvalue_checker::ByValueChecker, byvalue_scanner::identify_byvalue_safe_types,
-    type_database::TypeDatabase, types::make_ident, types::Namespace, types::TypeName,
+    byvalue_checker::ByValueChecker, conversion::ConvertError, type_database::TypeDatabase,
+    types::make_ident, types::Namespace, types::TypeName,
 };
 use proc_macro2::{TokenStream as TokenStream2, TokenTree};
 use quote::quote;
@@ -25,41 +25,16 @@ use syn::{
     ItemMod, ItemStruct, Type,
 };
 
-use super::{
+use super::super::{
     api::{Api, Use},
     bridge_name_tracker::BridgeNameTracker,
     codegen::{CodeGenerator, CodegenResults},
-    foreign_mod_converter::{ForeignModConversionCallbacks, ForeignModConverter},
     rust_name_tracker::RustNameTracker,
     type_converter::TypeConverter,
     utilities::generate_utilities,
 };
 
-#[derive(Debug)]
-pub enum ConvertError {
-    NoContent,
-    UnsafePODType(String),
-    UnexpectedForeignItem,
-    UnexpectedOuterItem,
-    UnexpectedItemInMod,
-    ComplexTypedefTarget(String),
-    UnexpectedThisType,
-}
-
-impl Display for ConvertError {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            ConvertError::NoContent => write!(f, "The initial run of 'bindgen' did not generate any content. This might be because none of the requested items for generation could be converted.")?,
-            ConvertError::UnsafePODType(err) => write!(f, "An item was requested using 'generate_pod' which was not safe to hold by value in Rust. {}", err)?,
-            ConvertError::UnexpectedForeignItem => write!(f, "Bindgen generated some unexpected code in a foreign mod section. You may have specified something in a 'generate' directive which is not currently compatible with autocxx.")?,
-            ConvertError::UnexpectedOuterItem => write!(f, "Bindgen generated some unexpected code in its outermost mod section. You may have specified something in a 'generate' directive which is not currently compatible with autocxx.")?,
-            ConvertError::UnexpectedItemInMod => write!(f, "Bindgen generated some unexpected code in an inner namespace mod. You may have specified something in a 'generate' directive which is not currently compatible with autocxx.")?,
-            ConvertError::ComplexTypedefTarget(ty) => write!(f, "autocxx was unable to produce a typdef pointing to the complex type {}.", ty)?,
-            ConvertError::UnexpectedThisType => write!(f, "Unexpected type for 'this'")?, // TODO give type/function
-        }
-        Ok(())
-    }
-}
+use super::parse_foreign_mod::{ForeignModConversionCallbacks, ForeignModConverter};
 
 #[derive(Copy, Clone, Eq, PartialEq)]
 enum TypeKind {
@@ -68,68 +43,10 @@ enum TypeKind {
     ForwardDeclaration, // no full C++ declaration available - can't even generate UniquePtr
 }
 
-/// Converts the bindings generated by bindgen into a form suitable
-/// for use with `cxx`.
-/// In fact, most of the actual operation happens within an
-/// individual `BridgeConversion`.
-///
-/// # Flexibility in handling bindgen output
-///
-/// autocxx is inevitably tied to the details of the bindgen output;
-/// e.g. the creation of a 'root' mod when namespaces are enabled.
-/// At the moment this crate takes the view that it's OK to panic
-/// if the bindgen output is not as expected. It may be in future that
-/// we need to be a bit more graceful, but for now, that's OK.
-pub(crate) struct BridgeConverter<'a> {
-    include_list: &'a [String],
-    type_database: &'a TypeDatabase,
-}
-
-impl<'a> BridgeConverter<'a> {
-    pub fn new(include_list: &'a [String], type_database: &'a TypeDatabase) -> Self {
-        Self {
-            include_list,
-            type_database,
-        }
-    }
-
-    /// Convert a TokenStream of bindgen-generated bindings to a form
-    /// suitable for cxx.
-    pub(crate) fn convert(
-        &mut self,
-        mut bindgen_mod: ItemMod,
-        exclude_utilities: bool,
-    ) -> Result<CodegenResults, ConvertError> {
-        match &mut bindgen_mod.content {
-            None => Err(ConvertError::NoContent),
-            Some((_, items)) => {
-                let items_to_process = items.drain(..).collect();
-                let items_in_root = find_items_in_root(items_to_process)?;
-                let byvalue_checker =
-                    identify_byvalue_safe_types(&items_in_root, &self.type_database)?;
-                let conversion = BridgeConversion {
-                    bindgen_mod,
-                    extern_c_mod: None,
-                    type_converter: TypeConverter::new(),
-                    byvalue_checker,
-                    include_list: &self.include_list,
-                    apis: Vec::new(),
-                    bridge_name_tracker: BridgeNameTracker::new(),
-                    rust_name_tracker: RustNameTracker::new(),
-                    type_database: &self.type_database,
-                    use_stmts_by_mod: HashMap::new(),
-                    incomplete_types: HashSet::new(),
-                };
-                conversion.convert_items(items_in_root, exclude_utilities)
-            }
-        }
-    }
-}
-
 /// A particular bridge conversion operation. This can really
 /// be thought of as a ton of parameters which we'd otherwise
 /// need to pass into each individual function within this file.
-struct BridgeConversion<'a> {
+pub(crate) struct BridgeConversion<'a> {
     bindgen_mod: ItemMod,
     extern_c_mod: Option<ItemForeignMod>,
     type_converter: TypeConverter,
@@ -143,25 +60,28 @@ struct BridgeConversion<'a> {
     incomplete_types: HashSet<TypeName>,
 }
 
-fn find_items_in_root(items: Vec<Item>) -> Result<Vec<Item>, ConvertError> {
-    for item in items {
-        match item {
-            Item::Mod(root_mod) => {
-                // With namespaces enabled, bindgen always puts everything
-                // in a mod called 'root'. We don't want to pass that
-                // onto cxx, so jump right into it.
-                assert!(root_mod.ident == "root");
-                if let Some((_, items)) = root_mod.content {
-                    return Ok(items);
-                }
-            }
-            _ => return Err(ConvertError::UnexpectedOuterItem),
+impl<'a> BridgeConversion<'a> {
+    pub(crate) fn new(
+        bindgen_mod: ItemMod,
+        byvalue_checker: ByValueChecker,
+        include_list: &'a [String],
+        type_database: &'a TypeDatabase,
+    ) -> Self {
+        BridgeConversion {
+            bindgen_mod,
+            extern_c_mod: None,
+            type_converter: TypeConverter::new(),
+            byvalue_checker,
+            include_list,
+            apis: Vec::new(),
+            bridge_name_tracker: BridgeNameTracker::new(),
+            rust_name_tracker: RustNameTracker::new(),
+            type_database,
+            use_stmts_by_mod: HashMap::new(),
+            incomplete_types: HashSet::new(),
         }
     }
-    Ok(Vec::new())
-}
 
-impl<'a> BridgeConversion<'a> {
     /// Main function which goes through and performs conversion from
     /// `bindgen`-style Rust output into `cxx::bridge`-style Rust input.
     /// At present, it significantly rewrites the bindgen mod,
@@ -176,7 +96,7 @@ impl<'a> BridgeConversion<'a> {
     /// One future option could be to provide a mode to bindgen where
     /// everything is opaque unless specifically allowlisted to be
     /// transparent.
-    fn convert_items(
+    pub(crate) fn convert_items(
         mut self,
         items: Vec<Item>,
         exclude_utilities: bool,
